+++ example/lib/const/resource.dart
@@ -1,27 +1,21 @@
-/// Generate by [resource_generator](https://github.com/CaiJingLong/flutter_resource_generator) library.
+/// Generate by [asset_generator](https://github.com/fluttercandies/flutter_asset_generator) library.
 /// PLEASE DO NOT EDIT MANUALLY.
+// ignore_for_file: constant_identifier_names
 class R {
   const R._();
 
-  /// ![preview](file:///Users/jinglongcai/code/dart/self/flutter_resource_generator/example/assets/bluetoothon-fjdfj.png)
   static const String ASSETS_BLUETOOTHON_FJDFJ_PNG =
       'assets/bluetoothon-fjdfj.png';
 
-  /// ![preview](file:///Users/jinglongcai/code/dart/self/flutter_resource_generator/example/assets/bluetoothon.png)
   static const String ASSETS_BLUETOOTHON_PNG = 'assets/bluetoothon.png';
 
-  /// ![preview](file:///Users/jinglongcai/code/dart/self/flutter_resource_generator/example/assets/camera.png)
   static const String ASSETS_CAMERA_PNG = 'assets/camera.png';
 
-  /// ![preview](file:///Users/jinglongcai/code/dart/self/flutter_resource_generator/example/images/audio.png)
   static const String IMAGES_AUDIO_PNG = 'images/audio.png';
 
-  /// ![preview](file:///Users/jinglongcai/code/dart/self/flutter_resource_generator/example/images/bluetoothoff.png)
   static const String IMAGES_BLUETOOTHOFF_PNG = 'images/bluetoothoff.png';
 
-  /// ![preview](file:///Users/jinglongcai/code/dart/self/flutter_resource_generator/example/images/child.png)
   static const String IMAGES_CHILD_PNG = 'images/child.png';
 
-  /// ![preview](file:///Users/jinglongcai/code/dart/self/flutter_resource_generator/example/images/course.png)
   static const String IMAGES_COURSE_PNG = 'images/course.png';
 }